Refactoring: separate concern of setting up Omise-PHP library out from Gateway\Http\Client\Payment.
Previously, this class was designed to be the only point to connect to Omise API. But, there will be more classes that need to connect with Omise API in the coming future and we can't just import 'Gateway\Http\Client\Payment' every time we need to make request to the API (it just doesn't make sense).

The benefit of this refactoring is, now 'Gateway\Http\Client\Payment.php' just taking care only 1 responsibility (placeRequest) and don't need to care of setting up Omise-PHP lib any more."
